Add `syntax` package
