--- src/main/kotlin/sigma/expressions/Term.kt
@@ -1,5 +1,0 @@
-package sigma.expressions
-
-abstract class Term {
-    abstract val location: SourceLocation
-}
