+++ src/main/kotlin/sigma/syntax/SourceLocation.kt
@@ -1,0 +1,19 @@
+package sigma.syntax
+
+import org.antlr.v4.runtime.ParserRuleContext
+
+data class SourceLocation(
+    val lineIndex: Int,
+    val columnIndex: Int,
+) {
+    companion object {
+        fun build(
+            ctx: ParserRuleContext,
+        ): SourceLocation = SourceLocation(
+            lineIndex = ctx.start.line,
+            columnIndex = ctx.start.charPositionInLine,
+        )
+    }
+
+    override fun toString(): String = "[Ln ${lineIndex}, Col ${columnIndex}]"
+}
