+++ src/main/kotlin/sigma/syntax/expressions/SymbolLiteral.kt
@@ -1,9 +1,10 @@
-package sigma.expressions
+package sigma.syntax.expressions
 
 
 import sigma.StaticTypeScope
 import sigma.StaticValueScope
 import sigma.parser.antlr.SigmaParser.SymbolLiteralAltContext
+import sigma.syntax.SourceLocation
 import sigma.types.SymbolType
 import sigma.types.Type
 import sigma.values.Symbol
@@ -15,11 +16,6 @@ data class SymbolLiteral(
     val symbol: Symbol,
 ) : Expression() {
     companion object {
-        fun of(name: String) = SymbolLiteral(
-            location = SourceLocation.Invalid,
-            symbol = Symbol.of(name),
-        )
-
         fun build(
             ctx: SymbolLiteralAltContext,
         ): SymbolLiteral = SymbolLiteral(
