+++ src/main/kotlin/sigma/syntax/expressions/Term.kt
@@ -1,0 +1,7 @@
+package sigma.syntax.expressions
+
+import sigma.syntax.SourceLocation
+
+abstract class Term {
+    abstract val location: SourceLocation
+}
