+++ src/test/kotlin/sigma/expressions/SymbolLiteralTests.kt
@@ -1,5 +1,6 @@
-package sigma.expressions
+package sigma.syntax.expressions
 
+import sigma.syntax.SourceLocation
 import sigma.values.Symbol
 import kotlin.test.Test
 import kotlin.test.assertEquals
